Funkcija za ucitavanje iz fajla
...dodata u Helpere. Funkkcija prima relativnu putanju iz storage, cita fajl i upisuje red po red u bazu podataka. Nema validacije za sada.

+++ database/migrations/2020_09_11_194145_create_products_table.php
@@ -21,7 +21,7 @@ class CreateProductsTable extends Migration
             $table->unsignedBigInteger('category');
             $table->foreign('category')->references('id')->on('categories')->onDelete('cascade');
             $table->char('manufacturor',50);
-            $table->integer('serial');
+            $table->bigInteger('serial');
             $table->integer('sku');
             $table->decimal('prise',10,2);
             $table->decimal('discount',10,2);
